Updated documentation and removed old code in various modules.

+++ modules/transport/Aggregation.h
@@ -75,8 +75,8 @@ private:
 
 
 	    // counters
-	    int nbAggrPktSentDown;
-	    int nbAggrPktReceived;
+	    long nbAggrPktSentDown;
+	    long nbAggrPktReceived;
 };
 
 #endif /* AGGREGATION_H_ */
